Rename apps files to something better
