+++ src/endpoints/apps/delete.rs
